refactoring test code and making enum equality comparison stable
