+++ src/test_cjt.py
@@ -6,43 +6,10 @@ from joinboost.cjt import CJT
 from joinboost.joingraph import JoinGraph
 from joinboost.semiring import AvgSemiRing, CountSemiRing, SumSemiRing
 from joinboost.aggregator import Annotation
+from src.test_utils import initialize_synthetic_one_to_many, initialize_synthetic_many_to_many
 
 
 class TestCJT(unittest.TestCase):
-    """
-    Join Graph for data/synthetic-many-to-many/
-    S(BE) - T(BF) - R(ABDH) 
-    """
-
-    def initialize_synthetic_many_to_many(self, semi_ring=AvgSemiRing()):
-        duck_db_conn = duckdb.connect(database=':memory:')
-        join_graph = JoinGraph(duck_db_conn)
-        cjt = CJT(semi_ring=semi_ring, join_graph=join_graph)
-        cjt.add_relation('R', attrs=["D", "H"],
-                         relation_address='../data/synthetic-many-to-many/R.csv')
-        cjt.add_relation('S', attrs=["E"],
-                         relation_address='../data/synthetic-many-to-many/S.csv')
-        cjt.add_relation('T', attrs=["F"],
-                         relation_address='../data/synthetic-many-to-many/T.csv')
-        cjt.add_join('R', 'S', ['B'], ['B']);
-        cjt.add_join('S', 'T', ['B'], ['B']);
-        return cjt
-
-    """
-    Join Graph for data/synthetic-one_to_many/
-    R(ABDH) - T(BFK) - S(FE)
-    """
-
-    def initialize_synthetic_one_to_many(self, semi_ring=AvgSemiRing()):
-        duck_db_conn = duckdb.connect(database=':memory:')
-        join_graph = JoinGraph(duck_db_conn)
-        cjt = CJT(semi_ring=semi_ring, join_graph=join_graph)
-        cjt.add_relation('R', attrs=["A", "D", "H"], relation_address='../data/synthetic-one-to-many/R.csv')
-        cjt.add_relation('S', attrs=["E"], relation_address='../data/synthetic-one-to-many/S.csv')
-        cjt.add_relation('T', relation_address='../data/synthetic-one-to-many/T.csv')
-        cjt.add_join('R', 'T', ['B'], ['B']);
-        cjt.add_join('S', 'T', ['F'], ['F']);
-        return cjt
 
     """
     Tests if message passing works in many to many join graph.
@@ -54,7 +21,7 @@ class TestCJT(unittest.TestCase):
     """
 
     def test_many_to_many(self):
-        cjt = self.initialize_synthetic_many_to_many(semi_ring=AvgSemiRing(relation='R', attr='A'))
+        cjt = initialize_synthetic_many_to_many(semi_ring=AvgSemiRing(relation='R', attr='A'))
         expected = cjt.exe.conn.execute(
             """
             SELECT SUM(A), count(*), R.B 
@@ -76,7 +43,7 @@ class TestCJT(unittest.TestCase):
     """
 
     def test_one_to_many(self):
-        cjt = self.initialize_synthetic_one_to_many(semi_ring=AvgSemiRing(relation='T', attr='K'))
+        cjt = initialize_synthetic_one_to_many(semi_ring=AvgSemiRing(relation='T', attr='K'))
         expected = cjt.exe.conn.execute(
             """
             SELECT SUM(T.K), count(*), T.B 
@@ -100,7 +67,7 @@ class TestCJT(unittest.TestCase):
     """
 
     def test_one_to_many_with_selection(self):
-        cjt = self.initialize_synthetic_one_to_many(semi_ring=AvgSemiRing(relation='T', attr='K'))
+        cjt = initialize_synthetic_one_to_many(semi_ring=AvgSemiRing(relation='T', attr='K'))
         expected = cjt.exe.conn.execute(
             """
             SELECT SUM(T.K), count(*), T.B 
@@ -126,7 +93,7 @@ class TestCJT(unittest.TestCase):
     """
 
     def test_many_to_many_with_selection(self):
-        cjt = self.initialize_synthetic_many_to_many(semi_ring=AvgSemiRing(relation='R', attr='A'))
+        cjt = initialize_synthetic_many_to_many(semi_ring=AvgSemiRing(relation='R', attr='A'))
         expected = cjt.exe.conn.execute(
             """
             SELECT SUM(A), count(*), R.B 
@@ -143,7 +110,7 @@ class TestCJT(unittest.TestCase):
 
     # TODO: add more tests for group-by
     def test_one_to_many_with_groupby(self):
-        cjt = self.initialize_synthetic_one_to_many(semi_ring=AvgSemiRing(relation='T', attr='K'))
+        cjt = initialize_synthetic_one_to_many(semi_ring=AvgSemiRing(relation='T', attr='K'))
         expected = cjt.exe.conn.execute(
             """
             SELECT SUM(T.K), count(*), T.B, S.E
@@ -158,7 +125,7 @@ class TestCJT(unittest.TestCase):
         self.assertEqual(expected, actual)
 
     def test_many_to_many_with_groupby(self):
-        cjt = self.initialize_synthetic_many_to_many(semi_ring=AvgSemiRing(relation='T', attr='F'))
+        cjt = initialize_synthetic_many_to_many(semi_ring=AvgSemiRing(relation='T', attr='F'))
         expected = cjt.exe.conn.execute(
             """
             SELECT SUM(T.F), count(*), T.B, S.E
@@ -173,7 +140,7 @@ class TestCJT(unittest.TestCase):
         self.assertEqual(expected, actual)
 
     def test_many_to_many_for_count_semiring(self):
-        cjt = self.initialize_synthetic_many_to_many(semi_ring=CountSemiRing(relation='T'))
+        cjt = initialize_synthetic_many_to_many(semi_ring=CountSemiRing(relation='T'))
         expected = cjt.exe.conn.execute(
             """
             SELECT count(*), T.B
@@ -187,7 +154,7 @@ class TestCJT(unittest.TestCase):
         self.assertEqual(expected, actual)
 
     def test_many_to_many_for_sum_semiring(self):
-        cjt = self.initialize_synthetic_many_to_many(semi_ring=SumSemiRing(relation='T', attr='F'))
+        cjt = initialize_synthetic_many_to_many(semi_ring=SumSemiRing(relation='T', attr='F'))
         expected = cjt.exe.conn.execute(
             """
             SELECT sum(T.F), T.B
@@ -199,32 +166,6 @@ class TestCJT(unittest.TestCase):
         cjt.calibration()
         actual = cjt.absorption('T', ['B'], ['B'], mode=3)
         self.assertEqual(expected, actual)
-
-    def test_multiplicity_for_many_to_many(self):
-        cjt = self.initialize_synthetic_many_to_many()
-
-        self.assertGreater(cjt.multiplicity['R']['S'], 1)
-        self.assertGreater(cjt.multiplicity['S']['R'], 1)
-        self.assertGreater(cjt.multiplicity['S']['T'], 1)
-        self.assertGreater(cjt.multiplicity['T']['S'], 1)
-
-        self.assertEqual(cjt.missing_keys['S']['R'], 1)
-        self.assertNotIn('S', cjt.missing_keys['R'])
-        self.assertNotIn('R', cjt.missing_keys['T'])
-        self.assertNotIn('T', cjt.missing_keys['R'])
-
-    def test_multiplicity_for_one_to_many(self):
-        cjt = self.initialize_synthetic_one_to_many()
-
-        self.assertGreater(cjt.multiplicity['R']['T'], 1)
-        self.assertEqual(cjt.multiplicity['T']['R'], 1)
-        self.assertGreater(cjt.multiplicity['S']['T'], 1)
-        self.assertEqual(cjt.multiplicity['T']['S'], 1)
-
-        self.assertEqual(cjt.missing_keys['S']['T'], 1)
-        self.assertNotIn('S', cjt.missing_keys['T'])
-        self.assertNotIn('R', cjt.missing_keys['T'])
-        self.assertNotIn('T', cjt.missing_keys['R'])
 
 
 if __name__ == '__main__':
